requirements.txt
faiss-gpu==1.7.1
joblib>=1.2.0
numpy>=1.22
Pillow>=9.0.1
pkg-resources==0.0.0
scikit-learn==0.24.2
scipy>=1.10.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.61.0
typing-extensions==3.10.0.0
pandas==2.0.3